GitFollowerBot
GitFollowerBot is a tool that automates logging into your GitHub account to fetch your follower list. Developed with Selenium, this bot allows users to specify the number of pages to quickly and accurately scrape follower data. It's an ideal solution for developers and anyone looking to manage their followers effortlessly.